feat: 优化任务调度Get请求接口参数处理方法
- 优化了任务查询接口，使用统一的 JobQuery 类
- 调整了分页参数类型，从 Integer 改为 Long
- 移除了未使用的 JobPageQuery 类
- 新增了通用的 PageQuery 类
- 简化了 JobController 和 JobLogController 的代码结构- 更新了 JobService 和 JobLogService 的实现

+++ vertx-app-service/vertx-framework/vertx-framework-extension/vertx-framework-extension-schedule/src/main/java/com/onezol/vertx/framework/extension/schedule/http/JobBatchHttpService.java
@@ -26,8 +26,8 @@ public interface JobBatchHttpService {
      */
     @GetExchange("/batch/list")
     JobPageResult<List<JobLog>> page(
-            @RequestParam("page") Integer page,
-            @RequestParam("size") Integer size,
+            @RequestParam("page") Long page,
+            @RequestParam("size") Long size,
             @RequestParam(value = "jobId", required = false) Long jobId,
             @RequestParam(value = "groupName", required = false) String groupName,
             @RequestParam(value = "jobName", required = false) String jobName,
